last commit

+++ modules/main/profile_controller.php
@@ -1,12 +1,15 @@
 <?php
-  if (isset($_GET['res'])){echo 'Update sucesfull';}
+  if (isset($_GET['res'])){echo PROFILE_UPDATE_OK;}
    
   function get_user_data(){
       $sql = "SELECT 
                    status
                         FROM users WHERE id ='".$_SESSION['user_id']."'";
-       $res = mysql_query($sql)  or die (mysql_error());       
-       $res = mysql_fetch_assoc($res);
+      // $res = mysql_query($sql)  or die (mysql_error());       
+      // $res = mysql_fetch_assoc($res);
+      $db = new PDO('mysql:host=localhost;dbname=test', 'root', 'vertrigo');
+       $column = $db->query($sql);
+       $res = $column->fetch(PDO::FETCH_ASSOC);
        
        $status = $res['status'];
       $sql =    "SELECT u.name,
@@ -27,18 +30,21 @@
                          previlegius p 
                     WHERE  u.id='".$_SESSION['user_id']."' AND p.status = '$status'";
          
-  
-       $res = mysql_query($sql)  or die (mysql_error());       
-       $res = mysql_fetch_assoc($res);
+       $column = $db->query($sql);
+       $res = $column->fetch(PDO::FETCH_ASSOC);
+     //  $res = mysql_query($sql)  or die (mysql_error());       
+     //  $res = mysql_fetch_assoc($res);
          return $res;
         
     }
+    
     $res = get_user_data();
    
     //echo '<pre>';
     //print_r($res);//exit;
      
   function  make_update($POST){
+      $db = new PDO('mysql:host=localhost;dbname=test', 'root', 'vertrigo');
     if (!empty($_FILES['uploadfile']['tmp_name'])){
          
      copy($_FILES['uploadfile']['tmp_name'],'skins/images/'.$_SESSION['user_id'].'.jpeg');
@@ -49,7 +55,9 @@
                             SET 
                               avatar  = '$avatar' 
                                 WHERE id='".$_SESSION['user_id']."'";
-            $res = mysql_query($sql)  or die (mysql_error());
+            $column = $db->query($sql);
+            $res = $column->fetch(PDO::FETCH_ASSOC);
+        // $res = mysql_query($sql)  or die (mysql_error());
       }
       extract($POST);
          $sql = "UPDATE users 
@@ -58,8 +66,9 @@
                               email = '$email',
                               pass  = '$pass' 
                                      WHERE id='".$_SESSION['user_id']."'";
-          
-        $res = mysql_query($sql)  or die (mysql_error());
+         $column = $db->query($sql);
+         $res = $column->fetch(PDO::FETCH_ASSOC); 
+      //  $res = mysql_query($sql)  or die (mysql_error());
       $_SESSION['user'] = $POST['name'];
       
              header('location:?type=profile&res=true');
@@ -67,27 +76,38 @@
     if (($POST['make_update'] == 1)) {make_update($POST);}
     
     function  make_dell(){
-      unlink('skins/images/'.$_SESSION['user_id'].'.jpeg'); 
+        $db = new PDO('mysql:host=localhost;dbname=test', 'root', 'vertrigo');
+        if (file_exists('skins/images/'.$_SESSION['user_id'].'.jpeg')){
+          unlink('skins/images/'.$_SESSION['user_id'].'.jpeg');
+       } 
         $sql = " DELETE FROM 
                             users
                                  WHERE id='".$_SESSION['user_id']."'";
-        $res = mysql_query($sql)  or die (mysql_error());
-        unlink('skins/images'.$_SESSION['user_id'].'.jpeg');
-         
+        // $column = $db->query($sql);
+       //  $res = $column->fetch(PDO::FETCH_ASSOC);
+      //$res = mysql_query($sql)  or die (mysql_error());
+        $del = $db->exec($sql);
+          
     }
-    if (($POST['make_dell'] == 1)) {make_dell();unset($_SESSION['user']);
+    if (($POST['make_dell'])) {
+                echo 'sdsds';make_dell();unset($_SESSION['user']);
       header('location:?type=preView');
     };
-    /*
-    $res = mysql_fetch_assoc($res);
-                      if ($res['pass'] !== $pass){
-                          $error_msg .= 'Password is incorect <br />';   
-                      } 
-   function res2array($data){
-            while($row = mysql_fetch_assoc($data)){
-                $arr[] = $row;
-            }
-            return $arr;  
-}   
-    */
-
+    
+     function show_profile (){
+         $db = new PDO('mysql:host=localhost;dbname=test', 'root', 'vertrigo');
+        $sql =    "SELECT  
+                            name,
+                            avatar,
+                            status,
+                            date_reg,
+                            date_last_log
+                            
+                     FROM users 
+                            WHERE id='".$_GET['profile_id']."'";
+         $column = $db->query($sql);
+        return $res = $column->fetch(PDO::FETCH_ASSOC); 
+    }
+    if(isset($_GET['profile_id'])){
+       $res = show_profile();
+    }
